requirements.txt
Flask==0.12
Flask-RESTful==0.3.5
msgpack-python==0.4.8
requests==2.12.4
